feat: 60kg 기준 운동별, 시간별 칼로리 소모량, 지방 소모량 계산해서 반환

+++ backend/src/main/java/com/ssafy/ssafit/domain/ExerciseLog.java
@@ -17,6 +17,6 @@ public class ExerciseLog {
     private Date exerciseDate;
     private String exerciseType;
     private int exerciseTime;
-    private Long caloriesBurned;
-    private Long fatBurned;
+    private double caloriesBurned;
+    private double fatBurned;
 }
